Add audio decorator
[what]
Add audio decorator which can make a tutorialView showing with audio.
And fix a hitTest bug.

[why]
Try to add new feature and maybe a tutorial with audio would be better.
HitTest bug is that some object of decorator which beneath top decorator can't
be touched.

[how]
Add a new decorator file.
In hitTest:event:, if there is no view should response then return nil.